Implemented changes for first assignment
Add verb and fruit support.  Add support for not repeating words if possible (I put in a short-circuit to prevent an infinite loop if the word lists are too short).  Now prints the number of replaced words.  (15 is correct for madtemplate.txt, 21 is correct for madtemplate2.txt, 6 is correct for madtemplate3.txt)